ASRC-52: using floats throughout the app
Risk Calculation results were previously using doubles, but we don't need that kind of
precision.

+++ srcalc/src/main/java/gov/va/med/srcalc/domain/model/DerivedTerm.java
@@ -74,7 +74,7 @@ public final class DerivedTerm extends ModelTerm
     }
 
     @Override
-    public double getSummand(final Map<Variable, Value> inputValues) throws MissingValuesException
+    public float getSummand(final Map<Variable, Value> inputValues) throws MissingValuesException
     {
         return fRule.apply(
                 new Rule.EvaluationContext(getCoefficient(), inputValues));
